Update matchups to be generated one character at a time to not overload server

+++ lib/controller/matchup_generation_controller.dart
@@ -2,12 +2,18 @@ import '../glory_backend.dart';
 
 class MatchupGenerationController extends HTTPController {
   @httpPost
-  Future<Response> generateMatchups() async {
-    var deleteQuery = new Query<Matchup>()..canModifyAllInstances = true;
-    await deleteQuery.delete();
-
+  Future<Response> generateMatchups(@HTTPPath("index") int index) async {
     var characterQuery = new Query<Character>()..returningProperties((c) => [c.serverName]);
     var characterNames = (await characterQuery.fetch()).map((c) => c.serverName).toList();
+
+    if (index >= characterNames.length) {
+      return new Response.notFound();
+    }
+
+    var name = characterNames[index];
+
+    var deleteQuery = new Query<Matchup>()..where.character = whereEqualTo(name);
+    await deleteQuery.delete();
 
     int savedMatchups = 0;
 
@@ -17,94 +23,75 @@ class MatchupGenerationController extends HTTPController {
 
     var allMatches = await matchesQuery.fetch();
 
-    for (var name in characterNames) {
-//      var matchesWithThisCharacterQuery = new Query<VGMatch>()
-//          ..where.participants.haveAtLeastOneWhere.characterName = name;
-//
-//      matchesWithThisCharacterQuery.joinMany((m) => m.participants)
-//        ..returningProperties((p) => [p.characterName, p.won]);
-//
-//      var matches = await matchesWithThisCharacterQuery.fetch();
+    var matches = allMatches.where((m) {
+      return (m.participants.firstWhere((p) => p.characterName == name, orElse: () => null) != null);
+    });
 
-      var matches = allMatches.where((m) {
-        return (m.participants.firstWhere((p) => p.characterName == name, orElse: () => null) != null);
-      });
+    int wonCount = 0, lostCount = 0;
 
-      int wonCount = 0, lostCount = 0;
-
-      matches.forEach((m) {
-        if (m.participants.firstWhere((p) => p.characterName == name).won) {
-          wonCount ++;
-        } else {
-          lostCount ++;
-        }
-      });
-
-      var selfMatchupQuery = new Query<Matchup>()
-        ..values.character = name
-        ..values.otherCharacter = name
-        ..values.gamesAgainst = 0
-        ..values.gamesAgainstWon = 0
-        ..values.gamesWith = (wonCount + lostCount)
-        ..values.gamesWithWon = wonCount;
-
-      await selfMatchupQuery.insert();
-      savedMatchups ++;
-
-      for (var otherName in characterNames) {
-        if (name == otherName) {
-          continue;
-        }
-
-//        var matchesWithBothCharactersQuery = new Query<VGMatch>()
-//          ..where.participants.haveAtLeastOneWhere.characterName = name
-//          ..where.participants.haveAtLeastOneWhere.characterName = otherName;
-//
-//        matchesWithBothCharactersQuery.joinMany((m) => m.participants)
-//          ..returningProperties((p) => [p.characterName, p.won]);
-//
-//        var pairMatches = await matchesWithBothCharactersQuery.fetch();
-
-        var pairMatches = allMatches.where((m) {
-          var character = m.participants.firstWhere((p) => p.characterName == name, orElse: () => null);
-          var otherCharacter = m.participants.firstWhere((p) => p.characterName == otherName, orElse: () => null);
-
-          return (character != null && otherCharacter != null);
-        });
-
-        int wonAgainstCount = 0, lostAgainstCount = 0, wonWithCount = 0, lostWithCount = 0;
-
-        pairMatches.forEach((m) {
-          var character = m.participants.firstWhere((p) => p.characterName == name);
-          var otherCharacter = m.participants.firstWhere((p) => p.characterName == otherName);
-
-
-          if (character.won == otherCharacter.won) {
-            if (character.won) {
-              wonWithCount ++;
-            } else {
-              lostWithCount ++;
-            }
-          } else {
-            if (character.won) {
-              wonAgainstCount ++;
-            } else {
-              lostAgainstCount ++;
-            }
-          }
-        });
-
-        var matchupQuery = new Query<Matchup>()
-          ..values.character = name
-          ..values.otherCharacter = otherName
-          ..values.gamesAgainst = (wonAgainstCount + lostAgainstCount)
-          ..values.gamesAgainstWon = wonAgainstCount
-          ..values.gamesWith = (wonWithCount + lostWithCount)
-          ..values.gamesWithWon = wonWithCount;
-
-        await matchupQuery.insert();
-        savedMatchups ++;
+    matches.forEach((m) {
+      if (m.participants.firstWhere((p) => p.characterName == name).won) {
+        wonCount ++;
+      } else {
+        lostCount ++;
       }
+    });
+
+    var selfMatchupQuery = new Query<Matchup>()
+      ..values.character = name
+      ..values.otherCharacter = name
+      ..values.gamesAgainst = 0
+      ..values.gamesAgainstWon = 0
+      ..values.gamesWith = (wonCount + lostCount)
+      ..values.gamesWithWon = wonCount;
+
+    await selfMatchupQuery.insert();
+    savedMatchups ++;
+
+    for (var otherName in characterNames) {
+      if (name == otherName) {
+        continue;
+      }
+
+      var pairMatches = allMatches.where((m) {
+        var character = m.participants.firstWhere((p) => p.characterName == name, orElse: () => null);
+        var otherCharacter = m.participants.firstWhere((p) => p.characterName == otherName, orElse: () => null);
+
+        return (character != null && otherCharacter != null);
+      });
+
+      int wonAgainstCount = 0, lostAgainstCount = 0, wonWithCount = 0, lostWithCount = 0;
+
+      pairMatches.forEach((m) {
+        var character = m.participants.firstWhere((p) => p.characterName == name);
+        var otherCharacter = m.participants.firstWhere((p) => p.characterName == otherName);
+
+
+        if (character.won == otherCharacter.won) {
+          if (character.won) {
+            wonWithCount ++;
+          } else {
+            lostWithCount ++;
+          }
+        } else {
+          if (character.won) {
+            wonAgainstCount ++;
+          } else {
+            lostAgainstCount ++;
+          }
+        }
+      });
+
+      var matchupQuery = new Query<Matchup>()
+        ..values.character = name
+        ..values.otherCharacter = otherName
+        ..values.gamesAgainst = (wonAgainstCount + lostAgainstCount)
+        ..values.gamesAgainstWon = wonAgainstCount
+        ..values.gamesWith = (wonWithCount + lostWithCount)
+        ..values.gamesWithWon = wonWithCount;
+
+      await matchupQuery.insert();
+      savedMatchups ++;
     }
 
     return new Response.ok("Saved ${savedMatchups} matchups");
